+ Déplacement de certains fichiers temporaires dans des répertoires par établissement (on se retrouvait avec 40000 fichiers dans un répertoire...). + Modification de l'import des identifiants de Gepi dans SACoche car les fichiers ont changé de nom dans Gepi.
git-svn-id: https://svn.devsesamath.net/svn/sacoche/trunk@297 3af1deee-4ace-4d6f-927a-55d690a2d960

+++ _inc/fonction_clean.php
@@ -204,16 +204,19 @@ function deleteBOM($file)
 
 function effacer_fichiers_temporaires($dossier,$nb_minutes)
 {
-	$date_limite = time() - $nb_minutes*60;
-	$tab_fichier = Lister_Contenu_Dossier($dossier);
-	foreach($tab_fichier as $fichier_nom)
+	if(is_dir($dossier))
 	{
-		$chemin_fichier = $dossier.'/'.$fichier_nom;
-		$extension = pathinfo($chemin_fichier,PATHINFO_EXTENSION);
-		$date_unix = filemtime($chemin_fichier);
-		if( (is_file($chemin_fichier)) && ($date_unix<$date_limite) && ($extension!='htm') )
+		$date_limite = time() - $nb_minutes*60;
+		$tab_fichier = Lister_Contenu_Dossier($dossier);
+		foreach($tab_fichier as $fichier_nom)
 		{
-			unlink($chemin_fichier);
+			$chemin_fichier = $dossier.'/'.$fichier_nom;
+			$extension = pathinfo($chemin_fichier,PATHINFO_EXTENSION);
+			$date_unix = filemtime($chemin_fichier);
+			if( (is_file($chemin_fichier)) && ($date_unix<$date_limite) && ($extension!='htm') )
+			{
+				unlink($chemin_fichier);
+			}
 		}
 	}
 }
